Fri, Aug 18, 2023, 11:58 AM +06:00

+++ beecrowd/1837.c
@@ -9,7 +9,7 @@ int main(void)
 
     //divident = divisor * quotient + reminder
     // a = b * q + r
-    q = a / (double)b;
+    q = a / b;
     r = a - (b * q);
 
     if(r < 0)
